requirements.txt
Django==4.2
djangorestframework==3.15.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
pre-commit==3.7.0
